+++ packages/feature-fetch/src/exceptions/index.ts
@@ -1,3 +1,3 @@
+export * from './FetchError';
 export * from './NetworkError';
 export * from './RequestError';
-export * from './ServiceError';