+++ packages/feature-fetch/src/helper/serializer/serialize-body/serialize-body-to-json.ts
@@ -1,10 +1,10 @@
-import { ServiceError } from '../../../exceptions';
+import { FetchError } from '../../../exceptions';
 
 export function serializeBodyToJson(body: unknown): string {
 	try {
 		return JSON.stringify(body);
 	} catch (error) {
-		throw new ServiceError('#ERR_SERIZALIZE_BODY', {
+		throw new FetchError('#ERR_SERIZALIZE_BODY', {
 			description: 'Failed to serialize body to JSON string!',
 			throwable: error instanceof Error ? error : undefined
 		});