+++ packages/google-webfonts-client/CHANGELOG.md
@@ -1,5 +1,12 @@
 # google-webfonts-client
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - feature-fetch@0.0.11
+
 ## 0.0.9
 
 ### Patch Changes